Switch to use ses-node-json-transformer
