+++ packages/api/src/period/utils/runCustomExportTransformer.ts
@@ -1,0 +1,12 @@
+import { transform } from 'ses-node-json-transform';
+import { ExportTransformer, PeriodDetailsGiverReceiverDto } from '../types';
+
+export const runCustomExportTransformer = (
+  data: PeriodDetailsGiverReceiverDto[],
+  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types, @typescript-eslint/no-explicit-any
+  context: any,
+  transformer: ExportTransformer
+): Object[] => {
+  const result = transform(data, transformer.map, context);
+  return result;
+};
